tests/submission/Microsoft/semantic/mark.html

<!doctype html>
<html>
 <head>
  <meta content="text/html; charset=utf-8" http-equiv="content-type" />
  <title>verify MARK is a HTMLElement</title>
  <script src="/resources/testharness.js"></script>
  <script src="http://www.w3c-test.org/resources/testharnessreport.js"></script>
 </head>
<body>
<div id="log"></div>
<p>This is just an example of what MARK can do:</p>
<pre>
 <code>
 var i: Integer;
 begin
  i := <mark id="mark">1.1</mark>;
 end.
 </code>
</pre>
<script type="text/javascript">
 test(function()
 {
    assert_array_equals( "[object HTMLElement]",
                         document.getElementById("mark").toString() )
 })
</script>
</body>
</html>
